refactor(mcp): split tool surface modules
