+++ crates/aether-mcp/src/tools/mod.rs
@@ -15,11 +15,6 @@
 //! populated by `load_component` / `replace_component`.
 
 use std::collections::{BTreeMap, HashMap};
-use std::future::Future;
-use std::io;
-use std::mem;
-use std::path::{Path, PathBuf};
-use std::pin::Pin;
 use std::sync::{Arc, Mutex};
 
 // `tokio::sync::Mutex` (the async one used by the per-engine refresh-
@@ -49,8 +44,8 @@ use aether_kinds::{
     TerminateEngineResult, UploadBinary, UploadBinaryResult, UploadComponent,
     UploadComponentResult,
     trace::{
-        DescribeTreeResult, DispatchTraced, DispatchTracedAck, MailNodeWire, TRACE_MAILBOX_NAME,
-        TraceTail, TraceTailResult,
+        DescribeTreeResult, DispatchTraced, MailNodeWire, TRACE_MAILBOX_NAME, TraceTail,
+        TraceTailResult,
     },
 };
 use base64::Engine as _;
@@ -80,9 +75,75 @@ use aether_kinds::{
 };
 use base64::engine::general_purpose::STANDARD;
 use std::time::Duration;
-use tokio::fs;
 use tokio::time;
 
+mod bytes;
+mod capture;
+mod components;
+mod envelope;
+mod ids;
+mod logs_cost;
+#[cfg(test)]
+mod loopback;
+mod render;
+mod reply;
+mod state;
+
+trait NamedMailSpec: Sync {
+    fn recipient_name(&self) -> &str;
+    fn kind_name(&self) -> &str;
+    fn params(&self) -> Option<&serde_json::Value>;
+}
+
+impl NamedMailSpec for CaptureMailSpec {
+    fn recipient_name(&self) -> &str {
+        &self.recipient_name
+    }
+
+    fn kind_name(&self) -> &str {
+        &self.kind_name
+    }
+
+    fn params(&self) -> Option<&serde_json::Value> {
+        self.params.as_ref()
+    }
+}
+
+impl NamedMailSpec for TracedMailSpec {
+    fn recipient_name(&self) -> &str {
+        &self.recipient_name
+    }
+
+    fn kind_name(&self) -> &str {
+        &self.kind_name
+    }
+
+    fn params(&self) -> Option<&serde_json::Value> {
+        self.params.as_ref()
+    }
+}
+
+#[cfg(test)]
+use self::bytes::{render_bytes_leaf_in, render_bytes_reply, resolve_bytes_params};
+use self::capture::capture_check;
+use self::components::{
+    component_config_bytes, components_all_loaded, reject_zero_replicas, replica_base_name,
+    replica_names, selector_with_explicit_export,
+};
+use self::envelope::{
+    engine_envelope, engine_envelope_by_id, local_envelope, validate_recipient_scope,
+};
+use self::ids::{
+    mail_id_to_json, parse_engine_id, parse_kind_id, parse_mailbox_id, resolve_handled_kind,
+    static_kind_name,
+};
+use self::logs_cost::{actor_logs_err_message, level_to_str, parse_level};
+#[cfg(test)]
+use self::loopback::{RouteInventorySink, RouteLoopbackConfig};
+use self::render::{
+    death_reason_parts, frame_size_aware_error, internal, internal_msg, json, render_shape,
+};
+use self::reply::{decode_reply_events, decode_traced_ack, strip_ack};
 /// Default wall-clock cap on `send_mail` / `send_mail_traced` awaiting a
 /// chain to settle (issue 1242). 300s — clears a provider cap's API
 /// timeout (the gemini cap's 180s, anthropic's 120s) with margin for
@@ -1352,607 +1413,6 @@ impl Mcp {
     }
 }
 
-impl Mcp {
-    /// Build one `MailSpec` into an `engine = Some` envelope and route
-    /// it through the hub, awaiting the substrate's terminal settle and
-    /// surfacing the correlated reply events (issue 1242). Returns the
-    /// collected reply envelopes plus a `timed_out` flag — the await is
-    /// bounded by [`AWAIT_TIMEOUT_DEFAULT_MS`] so a cap that never
-    /// replies returns at the cap rather than hanging.
-    async fn deliver_one(&self, spec: MailSpec) -> anyhow::Result<(Vec<MailEnvelope>, bool)> {
-        let envelope = self.build_mail_envelope(spec).await?;
-        let timeout = Duration::from_millis(u64::from(AWAIT_TIMEOUT_DEFAULT_MS));
-        self.session.call_collecting(envelope, timeout).await
-    }
-
-    /// [`Self::deliver_one`]'s fire-and-forget twin: build the envelope
-    /// and write the `Call` without awaiting any reply (issue 1242).
-    async fn deliver_one_fire(&self, spec: MailSpec) -> anyhow::Result<()> {
-        let envelope = self.build_mail_envelope(spec).await?;
-        self.session.fire(envelope).await
-    }
-
-    /// Resolve a component registry selector hub-local to its wasm bytes +
-    /// `@actor` export (ADR-0116). aether-mcp issues a `ResolveComponent` to
-    /// the `aether.engine` cap (no engine route — the store is hub-level),
-    /// which matches the selector to a single component and replies with the
-    /// wasm bytes from its store; aether-mcp then forwards those bytes to
-    /// the target substrate's `aether.component` mailbox. Shared by
-    /// `load_component`, `replace_component`, and the boot-manifest
-    /// pre-resolution, so the load seam stays path-free. An `Err` reply (no
-    /// match, or an attribute query matching more than one component) is a
-    /// clean tool error.
-    async fn resolve_component(&self, selector: &str) -> Result<ResolvedComponent, McpError> {
-        let reply = self
-            .session
-            .call_one(local_envelope(
-                ENGINE_CAP,
-                &ResolveComponent {
-                    selector: ComponentSelector {
-                        query: Some(selector.to_owned()),
-                        namespace: None,
-                        handled_kind: None,
-                    },
-                },
-            ))
-            .await
-            .map_err(|e| frame_size_aware_error(&format!("resolve_component {selector:?}"), e))?;
-        match ResolveComponentResult::decode_from_bytes(&reply.payload) {
-            Some(ResolveComponentResult::Ok {
-                wasm,
-                export,
-                manifest,
-                config_kind,
-                ..
-            }) => {
-                // ADR-0138: the bare-load entry is the manifest's opted-in
-                // `default_entry` (the single-actor namespace or the
-                // `export!(entry = …)` designation), NOT "the first actor
-                // by list position". A defaultless multi-actor module
-                // reports `None`, so replica-name derivation falls through
-                // to `name` / `export`.
-                let entry_namespace = manifest.default_entry;
-                Ok(ResolvedComponent {
-                    wasm,
-                    export,
-                    entry_namespace,
-                    config_kind,
-                })
-            }
-            Some(ResolveComponentResult::Err { error }) => Err(internal_msg(&error)),
-            None => Err(internal_msg("undecodable ResolveComponentResult")),
-        }
-    }
-
-    /// Pre-resolve a `spawn_substrate` boot list against the component
-    /// registry (ADR-0116) and stage it as a temp boot-manifest JSON the
-    /// hub injects as `AETHER_BOOT_MANIFEST` (issue 1776). For each spec
-    /// aether-mcp resolves the selector hub-local to its wasm bytes, writes
-    /// the bytes to a per-process-unique temp `.wasm`, and points the
-    /// manifest entry's `wasm` at that staged path — so the substrate boot
-    /// autoload path stays path-based, now fed by the registry rather than
-    /// host build paths. A `module@actor` selector's `@actor` half
-    /// populates the entry's `export` unless the spec set one explicitly.
-    /// Returns the staged paths so the caller cleans them all up once the
-    /// substrate has read them at boot.
-    async fn stage_boot_manifest(
-        &self,
-        components: &[ComponentSpec],
-    ) -> Result<StagedBootManifest, McpError> {
-        use std::env;
-        use std::process;
-        use std::sync::atomic::{AtomicU64, Ordering};
-        static SEQ: AtomicU64 = AtomicU64::new(0);
-
-        let mut wasm_paths: Vec<PathBuf> = Vec::with_capacity(components.len());
-        let mut config_paths: Vec<PathBuf> = Vec::new();
-        let mut entries: Vec<serde_json::Value> = Vec::with_capacity(components.len());
-        let mut expected_names: Vec<String> = Vec::with_capacity(components.len());
-        for spec in components {
-            reject_zero_replicas(spec.replicas, &spec.selector)?;
-            let resolve_selector =
-                selector_with_explicit_export(&spec.selector, spec.export.as_deref());
-            let resolved = self.resolve_component(&resolve_selector).await?;
-            let seq = SEQ.fetch_add(1, Ordering::Relaxed);
-            let wasm_path =
-                env::temp_dir().join(format!("aether-boot-wasm-{}-{seq}.wasm", process::id()));
-            fs::write(&wasm_path, &resolved.wasm).await.map_err(|e| {
-                internal_msg(&format!(
-                    "staging boot wasm for selector {resolve_selector:?}: {e}"
-                ))
-            })?;
-            let mut entry = serde_json::json!({ "wasm": wasm_path.to_string_lossy() });
-            if let Some(name) = &spec.name {
-                entry["name"] = serde_json::json!(name);
-            }
-            if let Some(config) = component_config_bytes(
-                resolved.config_kind.as_ref(),
-                spec.config.clone(),
-                spec.config_path.as_deref(),
-                &format!("spawn_substrate component {resolve_selector:?}"),
-            )
-            .await?
-            {
-                let seq = SEQ.fetch_add(1, Ordering::Relaxed);
-                let config_path =
-                    env::temp_dir().join(format!("aether-boot-config-{}-{seq}.bin", process::id()));
-                fs::write(&config_path, config).await.map_err(|e| {
-                    internal_msg(&format!(
-                        "staging boot config for selector {resolve_selector:?}: {e}"
-                    ))
-                })?;
-                entry["config"] = serde_json::json!(config_path.to_string_lossy());
-                config_paths.push(config_path);
-            }
-            // An explicit `export` wins over the selector's `@actor` half.
-            let export = spec.export.clone().or_else(|| resolved.export.clone());
-            if let Some(ref e) = export {
-                entry["export"] = serde_json::json!(e);
-            }
-            // issue 2626: pass `replicas` into the staged manifest entry
-            // verbatim — `autoload::expand_replicas` fans it out at read time.
-            if let Some(replicas) = spec.replicas {
-                entry["replicas"] = serde_json::json!(replicas);
-            }
-            // Derive the expected registered name(s) in the same precedence
-            // order the engine applies: caller-supplied name > export
-            // namespace > entry actor namespace. Fail loud if none is
-            // determinable: a spawn that can't name what it's waiting for is
-            // a bug to surface at stage time.
-            let ns = replica_base_name(
-                spec.name.as_deref(),
-                export.as_deref(),
-                resolved.entry_namespace.as_deref(),
-            )
-            .ok_or_else(|| {
-                internal_msg(&format!(
-                    "component {:?}: cannot determine expected registered name \
-                     (no `name`, `export`, or entry actor namespace in the wasm manifest); \
-                     set `name` or `export` on the ComponentSpec",
-                    spec.selector,
-                ))
-            })?;
-            match spec.replicas {
-                Some(replicas) => expected_names.extend(
-                    replica_names(&ns, replicas)
-                        .into_iter()
-                        .map(|name| format!("aether.component/aether.embedded:{name}")),
-                ),
-                None => expected_names.push(format!("aether.component/aether.embedded:{ns}")),
-            }
-            entries.push(entry);
-            wasm_paths.push(wasm_path);
-        }
-
-        let seq = SEQ.fetch_add(1, Ordering::Relaxed);
-        let manifest_path =
-            env::temp_dir().join(format!("aether-boot-manifest-{}-{seq}.json", process::id()));
-        let bytes = serde_json::to_vec(&serde_json::json!({ "components": entries }))
-            .map_err(|e| internal_msg(&format!("encoding boot manifest: {e}")))?;
-        fs::write(&manifest_path, bytes)
-            .await
-            .map_err(|e| internal_msg(&format!("staging boot manifest: {e}")))?;
-        Ok(StagedBootManifest {
-            manifest_path,
-            wasm_paths,
-            config_paths,
-            expected_names,
-        })
-    }
-
-    /// Resolve a `MailSpec` against the per-engine merged kind view
-    /// (static prefill + cached `ListKinds` reply, ADR-0091) and build
-    /// the `engine = Some` wire envelope — the shared front half of
-    /// [`Self::deliver_one`] / [`Self::deliver_one_fire`]. The resolve +
-    /// encode runs through [`Self::resolve_and_encode`], so an unknown-kind
-    /// miss *or* a field-mismatch encode failure (a kind widened in place,
-    /// ADR-0091 §3) triggers one `aether.inventory.kinds` refresh-and-retry
-    /// before the error surfaces.
-    async fn build_mail_envelope(&self, spec: MailSpec) -> anyhow::Result<MailEnvelope> {
-        // ADR-0098/0099 wire boundary: `recipient_name` is user-controlled
-        // and folds to a registry key, so cap its scope depth / byte size
-        // before it reaches `mailbox_id_from_path` (the fold itself stays
-        // infallible for static callers). A breach fails this mail item.
-        validate_recipient_scope(&spec.recipient_name)?;
-        let engine = EngineId(
-            Uuid::parse_str(&spec.engine_id)
-                .map_err(|e| anyhow::anyhow!("engine_id is not a valid UUID: {e}"))?,
-        );
-        let params = spec.params.unwrap_or(serde_json::Value::Null);
-        let (desc, payload) = self
-            .resolve_and_encode(engine, &spec.kind_name, params)
-            .await?;
-        Ok(MailEnvelope {
-            to: MailboxAddress {
-                engine: Some(engine),
-                // ADR-0099 §4: resolve the recipient by the parse → fold,
-                // so a `/`-rendered hosted / nested actor name
-                // (`aether.component/aether.component/aether.embedded:aether.camera`)
-                // reaches its lineage-folded id. A root-cap name is a
-                // single segment and folds to the same id `hash(name)`
-                // gives.
-                mailbox: mailbox_id_from_path(&spec.recipient_name),
-            },
-            from: None,
-            kind: KindId(kind_id_from_parts(&desc.name, &desc.schema)),
-            correlation_id: None,
-            payload,
-        })
-    }
-
-    /// Resolve `params` against the engine's live schema for `kind_name`
-    /// and schema-encode them, refreshing the per-engine kind cache once
-    /// on a staleness signal (ADR-0091 §3). The single shared encode path
-    /// behind `send_mail` ([`Self::build_mail_envelope`]),
-    /// `send_mail_traced` ([`Self::encode_traced_bundle`]), and
-    /// `capture_frame` ([`Self::encode_capture_bundle`]).
-    ///
-    /// Two staleness triggers refresh the cache from
-    /// `aether.inventory.kinds`, both bounded to exactly one refresh
-    /// round-trip:
-    ///
-    /// - **Unknown-kind miss** — handled one layer down by
-    ///   [`Self::lookup_descriptor`]: the name isn't in the cache, so the
-    ///   refresh runs before the descriptor is returned.
-    /// - **Field-mismatch encode failure** — the name *is* cached but its
-    ///   schema is stale (a kind widened in place, e.g. `aether.mouse_button`
-    ///   gaining a field). The name resolves, so `lookup_descriptor` returns
-    ///   a hit with the narrow schema, and `encode_schema` rejects the new
-    ///   field. That encode failure — and *only* an encode failure, not a
-    ///   `resolve_bytes_params` blob-IO error — triggers the same refresh,
-    ///   re-resolves + re-encodes against the fresh schema once, and returns.
-    ///
-    /// The retry re-runs `resolve_bytes_params` (not just `encode_schema`)
-    /// so a widening that adds a `Bytes` field descends into it and resolves
-    /// its `$`-sigil embed against the fresh schema. It is bounded to one
-    /// refresh + one re-encode and never re-enters the refresh branch, so
-    /// there is no retry loop by construction. If the fresh vocab still
-    /// rejects the params the retry's error surfaces; if the name vanished
-    /// from the fresh vocab (a removal, not a widening) the original encode
-    /// error surfaces.
-    async fn resolve_and_encode(
-        &self,
-        engine: EngineId,
-        kind_name: &str,
-        params: serde_json::Value,
-    ) -> anyhow::Result<(KindDescriptor, Vec<u8>)> {
-        let desc = self.lookup_descriptor(engine, kind_name).await?;
-        let resolved = resolve_bytes_params(params.clone(), &desc.schema, max_frame_size())
-            .await
-            .map_err(|e| anyhow::anyhow!("resolving blob params: {e}"))?;
-        match aether_codec::encode_schema(&resolved, &desc.schema) {
-            Ok(payload) => Ok((desc, payload)),
-            Err(encode_err) => {
-                // The cached schema may be stale from an in-place kind
-                // widening (ADR-0091 §3). Unlike the miss path, we don't
-                // re-check cache presence after taking the guard — a
-                // stale-schema name *is* present, so a presence re-check
-                // would wrongly short-circuit before refreshing. N
-                // concurrent field-mismatch retries can each issue a
-                // refresh RPC; an accepted cold-path cost (encode failures
-                // are mistakes, not hot traffic), still one refresh per
-                // retry, no loop.
-                let guard = self.refresh_guard(engine);
-                let _refresh = guard.lock().await;
-                self.refresh_engine_kinds(engine).await;
-                let Some(fresh) = self.cache_lookup(engine, kind_name) else {
-                    // The name vanished from the fresh vocab (a removal,
-                    // not a widening) — surface the original encode error.
-                    return Err(anyhow::anyhow!("param encode failed: {encode_err}"));
-                };
-                let resolved = resolve_bytes_params(params, &fresh.schema, max_frame_size())
-                    .await
-                    .map_err(|e| anyhow::anyhow!("resolving blob params: {e}"))?;
-                let payload = aether_codec::encode_schema(&resolved, &fresh.schema)
-                    .map_err(|e| anyhow::anyhow!("param encode failed: {e}"))?;
-                Ok((fresh, payload))
-            }
-        }
-    }
-
-    /// Ensure `engine`'s reverse-name map is built (ADR-0088 §8). On the
-    /// first need for an engine, fetch `aether.inventory.manifest` and
-    /// fold it into an [`EngineNames`]; on any subsequent call the cached
-    /// map is reused. An engine that doesn't answer the manifest (older
-    /// build, headless without the cap, transient error) gets an empty
-    /// map cached — every lookup then falls back to the hex tag rather
-    /// than erroring the tool or re-fetching on every render.
-    async fn ensure_names(&self, engine: EngineId) {
-        if self
-            .names
-            .lock()
-            .expect("reverse-name cache mutex is never poisoned")
-            .contains_key(&engine)
-        {
-            return;
-        }
-        // Fetch outside the lock — the await must not hold a std Mutex.
-        // No / undecodable reply caches an empty map so we fall back to
-        // hex for this engine without re-querying every render.
-        let manifest = self
-            .session
-            .call_one(engine_envelope(engine, INVENTORY_CAP, &Manifest {}))
-            .await
-            .ok()
-            .and_then(|reply| ManifestResult::decode_from_bytes(&reply.payload))
-            .unwrap_or_else(|| ManifestResult {
-                names: Vec::new(),
-                templates: Vec::new(),
-            });
-        let mut cache = self
-            .names
-            .lock()
-            .expect("reverse-name cache mutex is never poisoned");
-        // A concurrent session may have populated it while we awaited —
-        // first writer wins, both maps are equivalent.
-        cache
-            .entry(engine)
-            .or_insert_with(|| EngineNames::from_manifest(&manifest));
-    }
-
-    /// ADR-0091 §3 lookup → miss → refresh → retry → error flow. Look
-    /// up `kind_name` in the engine's encode cache; on a miss, fetch
-    /// the substrate's authoritative vocabulary via
-    /// `aether.inventory.kinds` and retry. The per-engine refresh is
-    /// collapsed under an async mutex so two concurrent misses on
-    /// different unknown names trigger one RPC, not two.
-    ///
-    /// Returns the matched descriptor; errors with `unknown kind: …`
-    /// after one refresh round-trip if the engine doesn't recognise
-    /// the name (a typoed kind, or a kind belonging to a component
-    /// that hasn't been loaded yet — distinguishable by the error
-    /// type at the substrate's later dispatch attempt).
-    ///
-    /// This covers only the unknown-**name** trigger. A cached name whose
-    /// *schema* has gone stale (a kind widened in place) still resolves to
-    /// a hit here; that field-mismatch staleness is caught one layer up in
-    /// [`Self::resolve_and_encode`], which refreshes on the encode failure.
-    async fn lookup_descriptor(
-        &self,
-        engine: EngineId,
-        kind_name: &str,
-    ) -> anyhow::Result<KindDescriptor> {
-        // Fast path: hit on the cache as it stands. `prefill_engine`
-        // populates the static `descriptors::all()` baseline on first
-        // touch so the very first send for a substrate-vocab kind
-        // doesn't trip a refresh.
-        self.prefill_engine(engine);
-        if let Some(desc) = self.cache_lookup(engine, kind_name) {
-            return Ok(desc);
-        }
-
-        // Miss: take the per-engine refresh mutex, then re-check (a
-        // concurrent waiter may have just refreshed) and refresh
-        // ourselves if still missing.
-        let guard = self.refresh_guard(engine);
-        let _refresh = guard.lock().await;
-        if let Some(desc) = self.cache_lookup(engine, kind_name) {
-            return Ok(desc);
-        }
-        self.refresh_engine_kinds(engine).await;
-        self.cache_lookup(engine, kind_name)
-            .ok_or_else(|| anyhow::anyhow!("unknown kind: {kind_name}"))
-    }
-
-    /// Seed `engine`'s cache from the substrate's static vocabulary
-    /// (`descriptors::all`) the first time the harness touches it. The
-    /// static set is process-global so a second engine with the same
-    /// build sees the same prefill, but the cache is keyed per-engine
-    /// because component-defined kinds aren't shared across engines.
-    #[allow(clippy::significant_drop_tightening)] // tight scope already
-    fn prefill_engine(&self, engine: EngineId) {
-        let mut cache = self
-            .kinds
-            .descriptors
-            .lock()
-            .expect("kinds-cache mutex is never poisoned");
-        cache.entry(engine).or_insert_with(|| {
-            descriptors::all()
-                .into_iter()
-                .map(|d| (d.name.clone(), d))
-                .collect()
-        });
-    }
-
-    /// Synchronous cache hit/miss check — no await, holds the std
-    /// `Mutex` only across the cloning lookup.
-    fn cache_lookup(&self, engine: EngineId, kind_name: &str) -> Option<KindDescriptor> {
-        let cache = self
-            .kinds
-            .descriptors
-            .lock()
-            .expect("kinds-cache mutex is never poisoned");
-        cache.get(&engine).and_then(|m| m.get(kind_name).cloned())
-    }
-
-    /// Fetch-or-create the per-engine refresh mutex. The mutex is
-    /// wrapped in an `Arc` so we can drop the cache lock before
-    /// awaiting; the only writers to `refresh_guards` are this
-    /// function itself, so it's a small concurrent-insert with no
-    /// upstream contention.
-    fn refresh_guard(&self, engine: EngineId) -> Arc<AsyncMutex<()>> {
-        let mut guards = self
-            .kinds
-            .refresh_guards
-            .lock()
-            .expect("kinds-cache refresh-guards mutex is never poisoned");
-        guards
-            .entry(engine)
-            .or_insert_with(|| Arc::new(AsyncMutex::new(())))
-            .clone()
-    }
-
-    /// Issue `ListKinds` against the engine's `aether.inventory`
-    /// mailbox and replace this engine's cache with the reply (folded
-    /// over the static prefill — a component's own kinds layer on top
-    /// of the substrate baseline, neither side wins exclusively).
-    /// A failed RPC / undecodable reply leaves the cache untouched so
-    /// the caller's retry surfaces the original "unknown kind" miss
-    /// rather than a transient RPC error.
-    async fn refresh_engine_kinds(&self, engine: EngineId) {
-        let Some(ListKindsResult { kinds }) = self
-            .session
-            .call_one(engine_envelope(engine, INVENTORY_CAP, &ListKinds {}))
-            .await
-            .ok()
-            .and_then(|reply| ListKindsResult::decode_from_bytes(&reply.payload))
-        else {
-            return;
-        };
-
-        // Decode each `schema_wire` back into a `SchemaType` via
-        // `wire::from_bytes`; an entry whose schema fails to decode is
-        // dropped (the substrate's wire form is canonical, so a decode
-        // failure is a substrate / aether-data version mismatch —
-        // better to skip the entry than panic the tool call).
-        let fresh: Vec<KindDescriptor> = kinds
-            .into_iter()
-            .filter_map(|wire| {
-                let schema = wire::from_bytes::<SchemaType>(&wire.schema_wire).ok()?;
-                Some(KindDescriptor {
-                    name: wire.name,
-                    schema,
-                })
-            })
-            .collect();
-
-        // Hold the cache lock only for the merge; the await above is
-        // already complete, so the `MutexGuard`'s significant `Drop`
-        // doesn't span any await point.
-        self.merge_into_engine_cache(engine, fresh);
-    }
-
-    /// Merge `fresh` into `engine`'s cache map, replacing any prior
-    /// entries with the same name. Factored out of `refresh_engine_kinds`
-    /// so the cache lock is acquired in a tight scope — no other state
-    /// hangs off the same critical section, and no await crosses the
-    /// guard.
-    #[allow(clippy::significant_drop_tightening)] // tight scope already
-    fn merge_into_engine_cache(&self, engine: EngineId, fresh: Vec<KindDescriptor>) {
-        let mut cache = self
-            .kinds
-            .descriptors
-            .lock()
-            .expect("kinds-cache mutex is never poisoned");
-        let map = cache.entry(engine).or_default();
-        for desc in fresh {
-            map.insert(desc.name.clone(), desc);
-        }
-    }
-
-    /// Snapshot the per-engine kind descriptor map for reply decoding
-    /// (issue 1804). Returns a clone of `engine`'s `name → descriptor`
-    /// map, or an empty map when the cache hasn't been seeded for that
-    /// engine yet. The engine cache is prefilled from the static substrate
-    /// vocabulary on the first `build_mail_envelope` call (`prefill_engine`),
-    /// so an empty snapshot only arises on paths that never encoded a kind
-    /// for this engine (e.g. a broken `engine_id` before `deliver_one` errored).
-    fn snapshot_engine_kinds(&self, engine: EngineId) -> HashMap<String, KindDescriptor> {
-        self.kinds
-            .descriptors
-            .lock()
-            .expect("kinds-cache mutex is never poisoned")
-            .get(&engine)
-            .cloned()
-            .unwrap_or_default()
-    }
-
-    /// Batch-resolve `ids` that the engine's static map missed via
-    /// `aether.inventory.resolve` and fold the answers (positive *and*
-    /// negative) into the per-engine dynamic cache. A no-op when `ids` is
-    /// empty or the resolve call fails — the unresolved ids then render
-    /// as hex tags. `ids` are raw `u64`s; the resolve wire takes tagged
-    /// strings, so each is encoded for the request and decoded back for
-    /// the cache key.
-    async fn resolve_dynamic(&self, engine: EngineId, ids: &[u64]) {
-        if ids.is_empty() {
-            return;
-        }
-        let tagged: Vec<String> = ids.iter().filter_map(|id| tagged_id::encode(*id)).collect();
-        if tagged.is_empty() {
-            return;
-        }
-        let Some(ResolveResult { resolved }) = self
-            .session
-            .call_one(engine_envelope(
-                engine,
-                INVENTORY_CAP,
-                &Resolve { ids: tagged },
-            ))
-            .await
-            .ok()
-            .and_then(|reply| ResolveResult::decode_from_bytes(&reply.payload))
-        else {
-            return;
-        };
-        let mut cache = self
-            .names
-            .lock()
-            .expect("reverse-name cache mutex is never poisoned");
-        if let Some(names) = cache.get_mut(&engine) {
-            for entry in resolved {
-                if let Ok(id) = tagged_id::decode(&entry.id) {
-                    names.cache_resolved(id, entry.name);
-                }
-            }
-        }
-    }
-
-    /// Reverse-render every id in a settled trace tree to a real name
-    /// (ADR-0088 §8). Builds the engine's reverse map if needed, collects
-    /// the ids that the static map misses, resolves them in one batched
-    /// `aether.inventory.resolve` query, then renders each `MailNodeWire`
-    /// through the now-populated map — falling back to the ADR-0064 hex
-    /// tag for any id that resolves to nothing. `Handle` ids stay hex
-    /// (they never enter the reverse map).
-    async fn render_mail_nodes(
-        &self,
-        engine: EngineId,
-        nodes: Vec<MailNodeWire>,
-    ) -> Vec<MailNodeJson> {
-        self.ensure_names(engine).await;
-
-        // Collect the mailbox / kind / thread ids that the static map
-        // misses, so one batched resolve covers the whole tree.
-        let mut misses: Vec<u64> = Vec::new();
-        {
-            let cache = self
-                .names
-                .lock()
-                .expect("reverse-name cache mutex is never poisoned");
-            if let Some(names) = cache.get(&engine) {
-                for node in &nodes {
-                    for id in node_reversible_ids(node) {
-                        if names.needs_resolve(id) {
-                            misses.push(id);
-                        }
-                    }
-                }
-            }
-        }
-        misses.sort_unstable();
-        misses.dedup();
-        self.resolve_dynamic(engine, &misses).await;
-
-        let cache = self
-            .names
-            .lock()
-            .expect("reverse-name cache mutex is never poisoned");
-        match cache.get(&engine) {
-            Some(names) => nodes
-                .into_iter()
-                .map(|node| mail_node_to_json(node, Some(names)))
-                .collect(),
-            // No map for this engine (shouldn't happen post-ensure, but
-            // be defensive): render every id as a hex tag.
-            None => nodes
-                .into_iter()
-                .map(|n| mail_node_to_json(n, None))
-                .collect(),
-        }
-    }
-}
-
 #[tool_handler]
 impl ServerHandler for Mcp {
     fn get_info(&self) -> ServerInfo {
@@ -1963,1270 +1423,6 @@ impl ServerHandler for Mcp {
         info.capabilities = ServerCapabilities::builder().enable_tools().build();
         info.server_info = server_info;
         info
-    }
-}
-
-/// Render a [`SchemaType`] as a one-line human-readable shape string —
-/// the compact form `describe_kinds` returns by default. The rendering is
-/// intentionally lossy (names only, not discriminants or `repr_c`) and is
-/// enough to build `send_mail` params for simple kinds without fetching
-/// the full schema. Depth is capped at 6 (`…` past that) per CLAUDE.md's
-/// recursion rule; schema depth is structurally bounded by the vocabulary
-/// but the cap is cheap insurance against pathological nesting.
-fn render_shape(ty: &SchemaType) -> String {
-    fn render(ty: &SchemaType, depth: u8) -> String {
-        if depth > 6 {
-            return "\u{2026}".to_owned();
-        }
-        match ty {
-            SchemaType::Unit => "{}".to_owned(),
-            SchemaType::Bool => "bool".to_owned(),
-            SchemaType::Scalar(p) => match p {
-                Primitive::U8 => "u8",
-                Primitive::U16 => "u16",
-                Primitive::U32 => "u32",
-                Primitive::U64 => "u64",
-                Primitive::I8 => "i8",
-                Primitive::I16 => "i16",
-                Primitive::I32 => "i32",
-                Primitive::I64 => "i64",
-                Primitive::F32 => "f32",
-                Primitive::F64 => "f64",
-            }
-            .to_owned(),
-            SchemaType::String => "String".to_owned(),
-            SchemaType::Bytes => "Bytes".to_owned(),
-            SchemaType::Option(inner) => format!("Option<{}>", render(inner, depth + 1)),
-            SchemaType::Vec(inner) => format!("Vec<{}>", render(inner, depth + 1)),
-            SchemaType::Array { element, len } => {
-                format!("[{}; {}]", render(element, depth + 1), len)
-            }
-            SchemaType::Struct { fields, .. } => {
-                let parts: Vec<String> = fields
-                    .iter()
-                    .map(|f| format!("{}: {}", f.name, render(&f.ty, depth + 1)))
-                    .collect();
-                format!("{{ {} }}", parts.join(", "))
-            }
-            SchemaType::Enum { variants } => {
-                let parts: Vec<String> = variants
-                    .iter()
-                    .map(|v| match v {
-                        EnumVariant::Unit { name, .. } => name.to_string(),
-                        EnumVariant::Tuple { name, fields, .. } => {
-                            let inner: Vec<String> =
-                                fields.iter().map(|f| render(f, depth + 1)).collect();
-                            format!("{}({})", name, inner.join(", "))
-                        }
-                        EnumVariant::Struct { name, fields, .. } => {
-                            let inner: Vec<String> = fields
-                                .iter()
-                                .map(|f| format!("{}: {}", f.name, render(&f.ty, depth + 1)))
-                                .collect();
-                            format!("{} {{ {} }}", name, inner.join(", "))
-                        }
-                    })
-                    .collect();
-                parts.join(" | ")
-            }
-            SchemaType::Map { key, value } => {
-                format!(
-                    "Map<{}, {}>",
-                    render(key, depth + 1),
-                    render(value, depth + 1)
-                )
-            }
-            SchemaType::TypeId(id) => {
-                if *id == MailboxId::TYPE_ID {
-                    "MailboxId".to_owned()
-                } else if *id == KindId::TYPE_ID {
-                    "KindId".to_owned()
-                } else {
-                    format!("TypeId({id:#x})")
-                }
-            }
-        }
-    }
-    render(ty, 0)
-}
-
-/// ADR-0098/0099 input hygiene: reject a `recipient_name` whose
-/// `/`-rendered scope path exceeds the depth or byte caps before it
-/// folds to a `MailboxId`. The MCP `send_mail` surface is the wire
-/// boundary for user-controlled names, so the aggregate-key guard lands
-/// here; [`aether_data::mailbox_id_from_path`] stays infallible for
-/// static callers.
-fn validate_recipient_scope(recipient_name: &str) -> anyhow::Result<()> {
-    let segments: Vec<&str> = recipient_name.split('/').collect();
-    validate_scope_path(&segments).map_err(|e| match e {
-        ScopePathError::TooDeep { limit } => {
-            anyhow::anyhow!("recipient_name has more than {limit} scope segments")
-        }
-        ScopePathError::TooLong { limit } => {
-            anyhow::anyhow!("recipient_name exceeds the {limit}-byte scope-path cap")
-        }
-    })
-}
-
-/// A component registry selector resolved to its bytes + `@actor` export
-/// (ADR-0116) — the front half of `load_component` / `replace_component` /
-/// the boot-manifest pre-resolution. `export` is the `module@actor`
-/// selector's actor half, threaded into the forwarded `LoadComponent.export`.
-/// `entry_namespace` is the first actor's `Actor::NAMESPACE` from the wasm
-/// manifest (per `export!` order), used by `stage_boot_manifest` to derive the
-/// expected registered name when neither `spec.name` nor an export is set.
-struct ResolvedComponent {
-    wasm: Vec<u8>,
-    export: Option<String>,
-    entry_namespace: Option<String>,
-    config_kind: Option<KindDescriptorWire>,
-}
-
-/// The temp files a `stage_boot_manifest` wrote (ADR-0116): the
-/// boot-manifest JSON the hub injects as `AETHER_BOOT_MANIFEST` plus the
-/// staged component `.wasm` files it points at. The substrate reads them
-/// at boot, before the spawn reply returns; the spawn caller
-/// [`cleanup`](StagedBootManifest::cleanup)s them once it has.
-/// `expected_names` carries the full `aether.component/aether.embedded:{ns}`
-/// lineage address computed for each spec, used by `spawn_substrate` to poll
-/// readiness by identity rather than count.
-struct StagedBootManifest {
-    manifest_path: PathBuf,
-    wasm_paths: Vec<PathBuf>,
-    config_paths: Vec<PathBuf>,
-    expected_names: Vec<String>,
-}
-
-impl StagedBootManifest {
-    /// Best-effort remove the staged manifest + every staged wasm file.
-    /// The substrate has already read them at boot by the time the spawn
-    /// reply returns, so a removal failure is harmless.
-    async fn cleanup(&self) {
-        let _ = fs::remove_file(&self.manifest_path).await;
-        for path in &self.wasm_paths {
-            let _ = fs::remove_file(path).await;
-        }
-        for path in &self.config_paths {
-            let _ = fs::remove_file(path).await;
-        }
-    }
-}
-
-/// Resolve an optional component init-config source (`config` inline JSON or
-/// `config_path` JSON file) and schema-encode it to the component's declared
-/// Config kind. No source returns `None`; a source for a no-config component is
-/// a tool error.
-async fn component_config_bytes(
-    config_kind: Option<&KindDescriptorWire>,
-    config: Option<serde_json::Value>,
-    config_path: Option<&str>,
-    context: &str,
-) -> Result<Option<Vec<u8>>, McpError> {
-    let value = match (config, config_path) {
-        (None, None) => return Ok(None),
-        (Some(_), Some(_)) => {
-            return Err(McpError::invalid_params(
-                format!("{context}: set only one of `config` or `config_path`"),
-                None,
-            ));
-        }
-        (Some(value), None) => value,
-        (None, Some(path)) => {
-            let bytes = fs::read(path).await.map_err(|e| {
-                McpError::invalid_params(
-                    format!("{context}: reading config_path {path:?}: {e}"),
-                    None,
-                )
-            })?;
-            serde_json::from_slice(&bytes).map_err(|e| {
-                McpError::invalid_params(
-                    format!("{context}: parsing config_path {path:?} as JSON: {e}"),
-                    None,
-                )
-            })?
-        }
-    };
-
-    let Some(config_kind) = config_kind else {
-        return Err(McpError::invalid_params(
-            format!(
-                "{context}: config JSON was provided but the component declares no Config kind"
-            ),
-            None,
-        ));
-    };
-    let schema = wire::from_bytes::<SchemaType>(&config_kind.schema_wire).map_err(|e| {
-        internal_msg(&format!(
-            "{context}: decoding config schema for {}: {e}",
-            config_kind.name
-        ))
-    })?;
-    let resolved = resolve_bytes_params(value, &schema, max_frame_size())
-        .await
-        .map_err(|e| {
-            McpError::invalid_params(
-                format!(
-                    "{context}: resolving config blob params for {}: {e}",
-                    config_kind.name
-                ),
-                None,
-            )
-        })?;
-    let bytes = aether_codec::encode_schema(&resolved, &schema).map_err(|e| {
-        McpError::invalid_params(
-            format!("{context}: config does not match {}: {e}", config_kind.name),
-            None,
-        )
-    })?;
-    Ok(Some(bytes))
-}
-
-/// Return `true` once every name in `want` is present in `actual`.
-/// Used by `wait_for_loaded_components` to drive identity-based readiness
-/// polling: the count variant (`actual.len() >= want.len()`) is insufficient
-/// because a baseline trampoline or an unrequested component can satisfy a
-/// count while a requested component is still absent.
-fn components_all_loaded(want: &[String], actual: &[String]) -> bool {
-    want.iter().all(|w| actual.iter().any(|a| a == w))
-}
-
-/// Fold an explicit `export` argument into the hub-local component resolve
-/// selector so the resolve reply's config descriptor matches the actor type
-/// that will instantiate. If the selector already carries `module@actor`, the
-/// explicit export wins by replacing the actor half.
-fn selector_with_explicit_export(selector: &str, export: Option<&str>) -> String {
-    let Some(export) = export else {
-        return selector.to_owned();
-    };
-    let module = selector
-        .split_once('@')
-        .map_or(selector, |(module, _)| module);
-    format!("{module}@{export}")
-}
-
-/// Resolve the base name a `replicas` fan-out derives its `{base}-{index}`
-/// names from (issue 2626), using the same precedence the component host
-/// itself applies at load: caller `name` > `export` > entry actor
-/// namespace. `None` when none of the three is available — the caller
-/// turns that into a clean tool error naming what to set. Shared by
-/// `stage_boot_manifest` (deriving `expected_names` to poll) and
-/// `load_component` (deriving each replica's load name), so both sides of
-/// a replicated load agree on what the components register as.
-fn replica_base_name(
-    name: Option<&str>,
-    export: Option<&str>,
-    entry_namespace: Option<&str>,
-) -> Option<String> {
-    name.or(export).or(entry_namespace).map(str::to_owned)
-}
-
-/// Derive the `{base}-{index}` name set a `replicas` fan-out registers
-/// under: every instance suffixed, no bare-name special case for index 0,
-/// so `replicas: 1` differs from an omitted field only by the `-0` suffix.
-fn replica_names(base: &str, replicas: u32) -> Vec<String> {
-    (0..replicas)
-        .map(|index| format!("{base}-{index}"))
-        .collect()
-}
-
-/// Reject `replicas: 0` (ADR-0090 §4 posture: a bad known value is a hard
-/// error, not a silent no-op) before it reaches any load dispatch.
-fn reject_zero_replicas(replicas: Option<u32>, selector: &str) -> Result<(), McpError> {
-    if replicas == Some(0) {
-        return Err(McpError::invalid_params(
-            format!("component {selector:?}: replicas must be at least 1 (got 0)"),
-            None,
-        ));
-    }
-    Ok(())
-}
-
-/// Build a `MailEnvelope` addressed at a hub-local mailbox
-/// (`engine = None`) carrying a typed kind.
-fn local_envelope<K: Kind>(mailbox: &str, kind: &K) -> MailEnvelope {
-    MailEnvelope {
-        to: MailboxAddress::local(mailbox_id_from_path(mailbox)),
-        from: None,
-        kind: K::ID,
-        correlation_id: None,
-        payload: kind.encode_into_bytes(),
-    }
-}
-
-/// Build a `MailEnvelope` addressed at a mailbox on a specific
-/// substrate (`engine = Some`) carrying a typed kind — the hub routes
-/// it through to that engine's proxy.
-fn engine_envelope<K: Kind>(engine: EngineId, mailbox: &str, kind: &K) -> MailEnvelope {
-    engine_envelope_by_id(engine, mailbox_id_from_path(mailbox), kind)
-}
-
-/// Like [`engine_envelope`] but addresses the recipient by
-/// [`MailboxId`] directly. The trace-tree guided walk (ADR-0086 Phase
-/// 3b) discovers recipients as ids embedded in `Sent` events, never as
-/// names — a `MailboxId` is a one-way name hash, so there's no name to
-/// reconstruct.
-fn engine_envelope_by_id<K: Kind>(engine: EngineId, mailbox: MailboxId, kind: &K) -> MailEnvelope {
-    MailEnvelope {
-        to: MailboxAddress {
-            engine: Some(engine),
-            mailbox,
-        },
-        from: None,
-        kind: K::ID,
-        correlation_id: None,
-        payload: kind.encode_into_bytes(),
-    }
-}
-
-/// Map a `capture_frame` check spec onto a wire [`FrameCheck`],
-/// resolving the reduction name. An unknown name is an invalid-params
-/// error so a typo aborts the capture cleanly before it reaches the
-/// wire (iamacoffeepot/aether#1777).
-fn capture_check(spec: &CaptureCheckSpec) -> Result<FrameCheck, McpError> {
-    let reduction = match spec.reduction.as_str() {
-        "not_all_black" => FrameReduction::NotAllBlack,
-        "differs_from_background" => FrameReduction::DiffersFromBackground,
-        "coverage" => FrameReduction::Coverage,
-        "centroid" => FrameReduction::Centroid,
-        "bounding_box" => FrameReduction::BoundingBox,
-        other => {
-            return Err(McpError::invalid_params(
-                format!(
-                    "capture_frame check: unknown reduction {other:?}; expected one of \
-                     not_all_black, differs_from_background, coverage, centroid, bounding_box"
-                ),
-                None,
-            ));
-        }
-    };
-    Ok(FrameCheck {
-        reduction,
-        tolerance: spec.tolerance,
-        background: spec.background,
-        region: spec.region.as_ref().map(|r| FrameRect {
-            min_x: r.min_x,
-            min_y: r.min_y,
-            max_x: r.max_x,
-            max_y: r.max_y,
-        }),
-    })
-}
-
-/// Parse a UUID-string `engine_id` (from `list_engines` /
-/// `spawn_substrate`) into an `EngineId`.
-fn parse_engine_id(s: &str) -> Result<EngineId, McpError> {
-    Uuid::parse_str(s)
-        .map(EngineId)
-        .map_err(|e| McpError::invalid_params(format!("engine_id is not a valid UUID: {e}"), None))
-}
-
-/// Parse a tagged mailbox-id string (`mbx-…`, ADR-0064) into a
-/// `MailboxId`.
-fn parse_mailbox_id(s: &str) -> Result<MailboxId, McpError> {
-    tagged_id::decode_with_tag(s, Tag::Mailbox)
-        .map(MailboxId)
-        .map_err(|e| McpError::invalid_params(format!("mailbox_id: {e}"), None))
-}
-
-/// Parse a kind-id string for the `actor_cost` filter: a tagged
-/// `knd-…` id (ADR-0064) or a raw decimal `u64`. The raw form is
-/// accepted because a cost row's id round-trips back through this
-/// filter and a caller may paste a non-tagged synthetic id.
-fn parse_kind_id(s: &str) -> Result<KindId, McpError> {
-    if let Ok(id) = tagged_id::decode_with_tag(s, Tag::Kind) {
-        return Ok(KindId(id));
-    }
-    s.parse::<u64>().map(KindId).map_err(|_| {
-        McpError::invalid_params(
-            format!("kind_id: not a tagged `knd-…` id or a decimal u64: {s:?}"),
-            None,
-        )
-    })
-}
-
-/// Resolve a `handled_kind` filter token (ADR-0116 `list_components`) to a
-/// [`KindId`]: a tagged `knd-…` id or a decimal `u64` resolves directly;
-/// otherwise the token is a kind name resolved against the static substrate
-/// vocabulary (`describe_kinds`'s source). An unknown name is an
-/// invalid-params error.
-fn resolve_handled_kind(s: &str) -> Result<KindId, McpError> {
-    if let Ok(id) = tagged_id::decode_with_tag(s, Tag::Kind) {
-        return Ok(KindId(id));
-    }
-    if let Ok(id) = s.parse::<u64>() {
-        return Ok(KindId(id));
-    }
-    descriptors::all()
-        .into_iter()
-        .find(|d| d.name == s)
-        .map(|d| KindId(kind_id_from_parts(&d.name, &d.schema)))
-        .ok_or_else(|| {
-            McpError::invalid_params(
-                format!("handled_kind: not a tagged `knd-…` id, a decimal u64, or a known kind name: {s:?}"),
-                None,
-            )
-        })
-}
-
-/// Best-effort resolve a [`KindId`] to its name from the static kind
-/// inventory the MCP harness ships with (`describe_kinds`'s source).
-/// Component-defined kinds aren't in the inventory and return `None`.
-/// Cold path — recomputes the inventory's ids on each call; the cost
-/// dump is a diagnostic, not a hot loop.
-fn static_kind_name(id: KindId) -> Option<String> {
-    descriptors::all()
-        .into_iter()
-        .find(|d| kind_id_from_parts(&d.name, &d.schema) == id.0)
-        .map(|d| d.name)
-}
-
-/// Transcode the correlated reply envelopes a `call_collecting` returned
-/// into the MCP wire shape (issue 1242). Per envelope: the tagged kind
-/// id, the best-effort kind name, and the best-effort `decode_schema` of
-/// the payload against the matching descriptor (`None` on an unknown kind
-/// or a decode miss). Decode priority (issue 1804): per-engine kind cache
-/// (`engine_kinds`, includes component-defined kinds populated from
-/// `ListKinds`) keyed by `declared_reply` (the handler's
-/// `HandlerCapability.reply` from ADR-0109) when it matches the envelope
-/// kind, then a general engine-cache scan for traced batches where the
-/// per-reply handler isn't known, then the static substrate vocabulary,
-/// then base64. On a clean decode the raw bytes are omitted (issue 1246).
-/// Order is preserved — arrival order.
-fn decode_reply_events(
-    envelopes: &[MailEnvelope],
-    engine_kinds: &HashMap<String, KindDescriptor>,
-    declared_reply: Option<KindId>,
-) -> Vec<ReplyEventJson> {
-    let static_descriptors = descriptors::all();
-    // Resolve the reply-`Bytes` spill threshold once for the whole batch
-    // (issue 2108) and thread it through every projection.
-    let inline_max = reply_inline_max_bytes();
-    envelopes
-        .iter()
-        .map(|env| {
-            // Resolve the schema for this reply envelope. Tier 1: engine
-            // cache targeted by the declared reply kind — this is the path
-            // that decodes component-defined reply kinds to `params` rather
-            // than base64. Tier 2: general engine-cache scan — covers
-            // send_mail_traced batches where `declared_reply` is None and
-            // any handler in the batch may have replied. Tier 3: static
-            // substrate vocabulary — the fallback for native chassis cap
-            // replies not yet in the engine cache. Tier 4: base64.
-            let descriptor: Option<KindDescriptor> = declared_reply
-                .filter(|&dr| dr == env.kind)
-                .and_then(|dr| {
-                    engine_kinds
-                        .values()
-                        .find(|d| kind_id_from_parts(&d.name, &d.schema) == dr.0)
-                        .cloned()
-                })
-                .or_else(|| {
-                    engine_kinds
-                        .values()
-                        .find(|d| kind_id_from_parts(&d.name, &d.schema) == env.kind.0)
-                        .cloned()
-                })
-                .or_else(|| {
-                    static_descriptors
-                        .iter()
-                        .find(|d| kind_id_from_parts(&d.name, &d.schema) == env.kind.0)
-                        .cloned()
-                });
-            let kind_name = descriptor.as_ref().map(|d| d.name.clone());
-            let (params, payload_bytes) = descriptor
-                .as_ref()
-                .and_then(|d| {
-                    // Render reply `Bytes` fields back to readable text /
-                    // base64 (issue 1944): the strict decoder emits a byte
-                    // array, the MCP front projects it for the caller.
-                    aether_codec::decode_schema(&env.payload, &d.schema)
-                        .ok()
-                        .map(|v| render_bytes_reply(v, &d.schema, inline_max))
-                })
-                .map_or_else(
-                    // Decode miss: base64 the raw payload as the fallback
-                    // (the only signal when `params` is `null`).
-                    || (None, Some(STANDARD.encode(&env.payload))),
-                    // Clean decode: `params` is the surfacing; omit the
-                    // raw bytes so they aren't duplicated as an int-array.
-                    |v| (Some(v), None),
-                );
-            ReplyEventJson {
-                // Render the kind id as the ADR-0064 tagged string the
-                // rest of the MCP wire uses, falling back to a hex
-                // literal on an unencodable (non-kind-domain) id.
-                kind_id: tagged_id::encode(env.kind.0)
-                    .unwrap_or_else(|| format!("{:#x}", env.kind.0)),
-                kind_name,
-                params,
-                payload_bytes,
-            }
-        })
-        .collect()
-}
-
-/// Decode the `DispatchTracedAck` from a `send_mail_traced` ack call's
-/// collected events (issue 1242). The synchronous ack is the *first*
-/// reply event the trace cap emits on the dispatch cid; later events are
-/// downstream cap replies handled separately. An absent or undecodable
-/// ack, or an `Err` ack, is a tool error.
-fn decode_traced_ack(events: &[MailEnvelope]) -> Result<MailId, McpError> {
-    let ack_env = events
-        .first()
-        .ok_or_else(|| internal_msg("send_mail_traced: no ack reply from the trace cap"))?;
-    let ack = DispatchTracedAck::decode_from_bytes(&ack_env.payload)
-        .ok_or_else(|| internal_msg("undecodable DispatchTracedAck"))?;
-    match ack {
-        DispatchTracedAck::Ok { root } => Ok(root),
-        DispatchTracedAck::Err { error } => Err(internal_msg(&format!(
-            "send_mail_traced dispatch failed: {error}"
-        ))),
-    }
-}
-
-/// The collected `send_mail_traced` events minus the leading ack (the
-/// `DispatchTracedAck` [`decode_traced_ack`] consumes), leaving the flat
-/// list of downstream cap replies to surface as `replies` (issue 1242).
-fn strip_ack(events: &[MailEnvelope]) -> &[MailEnvelope] {
-    events.get(1..).unwrap_or(&[])
-}
-impl Mcp {
-    /// Encode a `send_mail_traced` batch into the same `NamedMail`
-    /// shape `CaptureFrame` carries: name-level addressing + schema-
-    /// encoded payload. The substrate's `TraceObserver` resolves the
-    /// names through its registry at dispatch time. Same lookup path
-    /// `encode_capture_bundle` uses (per-engine merged view, ADR-0091),
-    /// just over `TracedMailSpec` instead of `CaptureMailSpec`.
-    async fn encode_traced_bundle(
-        &self,
-        engine: EngineId,
-        specs: &[TracedMailSpec],
-    ) -> anyhow::Result<Vec<NamedMail>> {
-        let mut out = Vec::with_capacity(specs.len());
-        for spec in specs {
-            let params = spec.params.clone().unwrap_or(serde_json::Value::Null);
-            let (_desc, payload) = self
-                .resolve_and_encode(engine, &spec.kind_name, params)
-                .await
-                .map_err(|e| anyhow::anyhow!("{e} (kind {})", spec.kind_name))?;
-            out.push(NamedMail {
-                recipient_name: spec.recipient_name.clone(),
-                kind_name: spec.kind_name.clone(),
-                payload,
-                count: 1,
-            });
-        }
-        Ok(out)
-    }
-}
-
-/// Render a raw `u64` mailbox / kind / thread id to its display string
-/// (ADR-0088 §8): the engine's real name when `names` resolves it, else
-/// the ADR-0064 tagged-id string (`mbx-…` / `knd-…` / `thr-…`), else a
-/// hex literal if the tag bits are unencodable. `names == None` (no
-/// reverse map for the engine) renders the tag directly — the unchanged
-/// pre-inventory output.
-fn render_id(id: u64, names: Option<&EngineNames>) -> String {
-    names.map_or_else(
-        || tagged_id::encode(id).unwrap_or_else(|| format!("{id:#x}")),
-        |names| names.render(id),
-    )
-}
-
-/// Reverse-render a [`MailboxId`] through the engine's name map (or the
-/// hex tag on a miss / no map). Chassis-minted ids always carry tag bits,
-/// so the hex fallback never reaches the `{:#x}` arm in practice.
-fn mailbox_id_to_tagged(id: MailboxId, names: Option<&EngineNames>) -> String {
-    render_id(id.0, names)
-}
-
-fn kind_id_to_tagged(id: KindId, names: Option<&EngineNames>) -> String {
-    render_id(id.0, names)
-}
-
-fn mail_id_to_json(id: MailId, names: Option<&EngineNames>) -> MailIdJson {
-    MailIdJson {
-        sender: mailbox_id_to_tagged(id.sender, names),
-        correlation_id: id.correlation_id,
-    }
-}
-
-fn mail_node_to_json(node: MailNodeWire, names: Option<&EngineNames>) -> MailNodeJson {
-    MailNodeJson {
-        mail_id: mail_id_to_json(node.mail_id, names),
-        parent: node.parent.map(|p| mail_id_to_json(p, names)),
-        sender: mailbox_id_to_tagged(node.sender, names),
-        recipient: mailbox_id_to_tagged(node.recipient, names),
-        kind: kind_id_to_tagged(node.kind, names),
-        t_construct_start: node.t_construct_start.0,
-        t_sent: node.t_sent.0,
-        t_received: node.t_received.map(|n| n.0),
-        t_finished: node.t_finished.map(|n| n.0),
-        thread_name: node.thread_name,
-    }
-}
-
-/// The mailbox / kind / thread ids in one `MailNodeWire` that reverse
-/// through the inventory (ADR-0088 §8): the two mailbox endpoints, the
-/// kind, and both `MailId` senders. `correlation_id` is a `Uuid`, not a
-/// tagged id, so it's excluded. Thread ids ride in `thread_name` already
-/// resolved substrate-side, so they aren't re-resolved here.
-fn node_reversible_ids(node: &MailNodeWire) -> Vec<u64> {
-    let mut ids = vec![
-        node.sender.0,
-        node.recipient.0,
-        node.kind.0,
-        node.mail_id.sender.0,
-    ];
-    if let Some(parent) = &node.parent {
-        ids.push(parent.sender.0);
-    }
-    ids
-}
-
-impl Mcp {
-    /// Encode a `capture_frame` mail bundle: resolve each spec's kind
-    /// against the per-engine merged view (ADR-0091, static prefill +
-    /// cached `ListKinds` reply), schema-encode its params, and wrap
-    /// into the substrate-side `aether_kinds::NamedMail` shape
-    /// (name-level addressing + pre-encoded payload).
-    async fn encode_capture_bundle(
-        &self,
-        engine: EngineId,
-        specs: &[CaptureMailSpec],
-    ) -> anyhow::Result<Vec<NamedMail>> {
-        let mut out = Vec::with_capacity(specs.len());
-        for spec in specs {
-            let params = spec.params.clone().unwrap_or(serde_json::Value::Null);
-            let (_desc, payload) = self
-                .resolve_and_encode(engine, &spec.kind_name, params)
-                .await
-                .map_err(|e| anyhow::anyhow!("{e} (kind {})", spec.kind_name))?;
-            out.push(NamedMail {
-                recipient_name: spec.recipient_name.clone(),
-                kind_name: spec.kind_name.clone(),
-                payload,
-                count: 1,
-            });
-        }
-        Ok(out)
-    }
-}
-
-/// Serialize a tool result to the JSON string `rmcp` wraps as text
-/// content.
-fn json<T: serde::Serialize>(value: &T) -> Result<String, McpError> {
-    serde_json::to_string(value).map_err(|e| McpError::internal_error(e.to_string(), None))
-}
-
-/// Flatten a wire [`DeathReason`] into the `(reason, detail)` pair the
-/// `list_engines` tool renders: a short tag plus the variant's detail
-/// string (empty for the clean `Terminated` case). Flat over a tagged
-/// JSON enum so an LLM consumer reads the cause without a nested match.
-fn death_reason_parts(reason: DeathReason) -> (String, String) {
-    match reason {
-        DeathReason::Terminated => ("terminated".to_owned(), String::new()),
-        DeathReason::Crashed { detail } => ("crashed".to_owned(), detail),
-        DeathReason::Evicted { detail } => ("evicted".to_owned(), detail),
-        DeathReason::SpawnFailed { detail } => ("spawn_failed".to_owned(), detail),
-    }
-}
-
-// `e` is owned because callers do `.map_err(internal)` — the closure-
-// converted form needs an `FnOnce(anyhow::Error) -> McpError`.
-#[allow(clippy::needless_pass_by_value)]
-fn internal(e: anyhow::Error) -> McpError {
-    McpError::internal_error(e.to_string(), None)
-}
-
-fn internal_msg(msg: &str) -> McpError {
-    McpError::internal_error(msg.to_owned(), None)
-}
-
-/// iamacoffeepot/aether#1271: tools that ship potentially-large
-/// payloads through the RPC framing (currently `load_component` /
-/// `replace_component`) surface a `FrameTooLarge` / `EncodeTooLarge`
-/// failure as `invalid_params` rather than `internal_error`. The
-/// payload is a client-controllable input (the user picked the wasm
-/// path), and the actionable remediation — build the release wasm,
-/// raise `AETHER_MAX_FRAME_SIZE` — is specific to the caller. Falls
-/// through to `internal` for every other shape.
-///
-/// Detection is by substring of the error chain because the structured
-/// `RpcError` rides under `anyhow::Error` (the session's `call_once`
-/// formats the wire error with `{e:?}` into a string; the encode-side
-/// classifier formats `RpcClientError::Frame(...)` with `{e}`). Both
-/// shapes embed the literal `frame too large` / `encoded frame too
-/// large` strings the codec / RPC error variants produce.
-#[allow(clippy::needless_pass_by_value)]
-fn frame_size_aware_error(context: &str, e: anyhow::Error) -> McpError {
-    let text = e.to_string();
-    if text.contains("frame too large")
-        || text.contains("encoded frame too large")
-        || text.contains("FrameTooLarge")
-        || text.contains("EncodeTooLarge")
-    {
-        return McpError::invalid_params(
-            format!(
-                "{context}: payload exceeds the RPC framing cap — typically because the supplied \
-                 wasm is a debug build. Build the release wasm (target/wasm32-unknown-unknown/\
-                 release/*.wasm) or raise the cap via the AETHER_MAX_FRAME_SIZE env var. \
-                 Underlying: {text}",
-            ),
-            None,
-        );
-    }
-    McpError::internal_error(text, None)
-}
-
-/// Blob-embed preprocessor (issue 1944). The wire codec is strict and
-/// canonical — a `SchemaType::Bytes` param encodes only from a JSON byte
-/// array — so the consumer-facing ergonomics live here, in the MCP front
-/// that already owns the JSON params before schema-encoding them. Walk
-/// `value` alongside `schema` and, at every `Bytes` node, resolve a
-/// `$`-sigil embed object into the canonical byte array `encode_schema`
-/// accepts. A literal `[…]` array passes straight through (back-compat);
-/// a one-key embed object expands — `{"$file": path}` reads the file on
-/// the harness host and inlines its bytes, `{"$base64": s}` decodes,
-/// `{"$text": s}` UTF-8-encodes. Any other shape at a `Bytes` node, or an
-/// unknown `$`-tag, errors. Recursion depth is bounded by the
-/// compile-time kind schema, not by user-controlled runtime data, so it
-/// mirrors `encode_schema`'s own recursive walk; only the arms that can
-/// carry a `Bytes` leaf (`Struct` / `Option` / `Vec` / `Array` / `Map`)
-/// descend, every other value passes through untouched. `max_file_bytes`
-/// is the RPC frame cap a `{"$file"}` read is guarded against; the
-/// production call sites pass `max_frame_size()`.
-#[allow(clippy::too_many_lines)] // one arm per SchemaType variant; splitting the enum arm out would obscure the symmetry with render_bytes_reply
-fn resolve_bytes_params<'a>(
-    value: serde_json::Value,
-    schema: &'a SchemaType,
-    max_file_bytes: usize,
-) -> Pin<Box<dyn Future<Output = anyhow::Result<serde_json::Value>> + Send + 'a>> {
-    use serde_json::Value;
-    Box::pin(async move {
-        match schema {
-            SchemaType::Bytes => resolve_bytes_embed(value, max_file_bytes).await,
-            SchemaType::Option(inner) => {
-                if value.is_null() {
-                    Ok(value)
-                } else {
-                    resolve_bytes_params(value, inner, max_file_bytes).await
-                }
-            }
-            SchemaType::Vec(inner) => match value {
-                Value::Array(items) => {
-                    let mut out = Vec::with_capacity(items.len());
-                    for item in items {
-                        out.push(resolve_bytes_params(item, inner, max_file_bytes).await?);
-                    }
-                    Ok(Value::Array(out))
-                }
-                other => Ok(other),
-            },
-            SchemaType::Array { element, .. } => match value {
-                Value::Array(items) => {
-                    let mut out = Vec::with_capacity(items.len());
-                    for item in items {
-                        out.push(resolve_bytes_params(item, element, max_file_bytes).await?);
-                    }
-                    Ok(Value::Array(out))
-                }
-                other => Ok(other),
-            },
-            SchemaType::Struct { fields, .. } => match value {
-                Value::Object(mut map) => {
-                    for field in fields.iter() {
-                        if let Some(slot) = map.remove(&*field.name) {
-                            let resolved =
-                                resolve_bytes_params(slot, &field.ty, max_file_bytes).await?;
-                            map.insert(field.name.to_string(), resolved);
-                        }
-                    }
-                    Ok(Value::Object(map))
-                }
-                other => Ok(other),
-            },
-            SchemaType::Map {
-                value: value_schema,
-                ..
-            } => match value {
-                Value::Object(mut map) => {
-                    let keys: Vec<String> = map.keys().cloned().collect();
-                    for key in keys {
-                        if let Some(slot) = map.remove(&key) {
-                            let resolved =
-                                resolve_bytes_params(slot, value_schema, max_file_bytes).await?;
-                            map.insert(key, resolved);
-                        }
-                    }
-                    Ok(Value::Object(map))
-                }
-                other => Ok(other),
-            },
-            SchemaType::Enum { variants } => match value {
-                // Unit variant — bare string, no payload to descend into.
-                Value::String(_) => Ok(value),
-                // Non-unit variant — externally-tagged one-key object
-                // `{"VariantName": payload}`. Resolve the tag, then recurse
-                // into the payload according to the matched variant kind.
-                Value::Object(map) if map.len() == 1 => {
-                    let (tag, payload) = map.into_iter().next().expect("len == 1");
-                    let variant = variants.iter().find(|v| v.name() == tag.as_str());
-                    let resolved = match variant {
-                        None | Some(EnumVariant::Unit { .. }) => payload,
-                        Some(EnumVariant::Tuple { fields, .. }) => match fields.as_ref() {
-                            [single] => {
-                                resolve_bytes_params(payload, single, max_file_bytes).await?
-                            }
-                            _ => match payload {
-                                Value::Array(items) => {
-                                    let mut out = Vec::with_capacity(items.len());
-                                    for (item, field) in items.into_iter().zip(fields.iter()) {
-                                        out.push(
-                                            resolve_bytes_params(item, field, max_file_bytes)
-                                                .await?,
-                                        );
-                                    }
-                                    Value::Array(out)
-                                }
-                                other => other,
-                            },
-                        },
-                        Some(EnumVariant::Struct { fields, .. }) => match payload {
-                            Value::Object(mut inner) => {
-                                for field in fields.iter() {
-                                    if let Some(slot) = inner.remove(&*field.name) {
-                                        let resolved =
-                                            resolve_bytes_params(slot, &field.ty, max_file_bytes)
-                                                .await?;
-                                        inner.insert(field.name.to_string(), resolved);
-                                    }
-                                }
-                                Value::Object(inner)
-                            }
-                            other => other,
-                        },
-                    };
-                    let mut out = serde_json::Map::new();
-                    out.insert(tag, resolved);
-                    Ok(Value::Object(out))
-                }
-                other => Ok(other),
-            },
-            // Scalars, String, TypeId, Unit, Bool: no `Bytes` leaf is
-            // reachable through the embed grammar, so pass through.
-            _ => Ok(value),
-        }
-    })
-}
-
-/// Resolve a single `Bytes`-node value into the canonical JSON byte
-/// array (issue 1944). A literal `[…]` array passes through; a one-key
-/// `$`-sigil object expands into bytes; anything else errors. `{"$file"}`
-/// is guarded above `max_file_bytes` (the RPC frame cap) so a blob too
-/// large to ride in mail errors with a pointer to the staged-path
-/// mechanism rather than being silently inlined.
-async fn resolve_bytes_embed(
-    value: serde_json::Value,
-    max_file_bytes: usize,
-) -> anyhow::Result<serde_json::Value> {
-    use serde_json::Value;
-    let obj = match value {
-        // Canonical form — already a byte array. Back-compat passthrough.
-        Value::Array(_) => return Ok(value),
-        Value::Object(map) => map,
-        _ => anyhow::bail!(
-            "a Bytes field accepts a byte array or a single $-sigil embed \
-             ($file / $base64 / $text)"
-        ),
-    };
-    if obj.len() != 1 {
-        anyhow::bail!(
-            "a Bytes embed object must have exactly one $-sigil key \
-             ($file / $base64 / $text)"
-        );
-    }
-    let (key, body) = obj.into_iter().next().expect("len == 1");
-    let bytes: Vec<u8> = match key.as_str() {
-        "$file" => {
-            let path = body
-                .as_str()
-                .ok_or_else(|| anyhow::anyhow!("$file value must be a string path"))?;
-            let bytes = fs::read(path)
-                .await
-                .map_err(|e| anyhow::anyhow!("$file: reading {path:?}: {e}"))?;
-            if bytes.len() > max_file_bytes {
-                anyhow::bail!(
-                    "$file {path:?} is {} bytes, over the {max_file_bytes}-byte RPC frame cap; a \
-                     blob this large must stage as a hub-read path (ADR-0115/0116), not inline \
-                     into mail",
-                    bytes.len()
-                );
-            }
-            bytes
-        }
-        "$base64" => {
-            let s = body
-                .as_str()
-                .ok_or_else(|| anyhow::anyhow!("$base64 value must be a string"))?;
-            STANDARD
-                .decode(s)
-                .map_err(|e| anyhow::anyhow!("$base64: invalid base64: {e}"))?
-        }
-        "$text" => {
-            let s = body
-                .as_str()
-                .ok_or_else(|| anyhow::anyhow!("$text value must be a string"))?;
-            s.as_bytes().to_vec()
-        }
-        other => anyhow::bail!(
-            "a Bytes field accepts a byte array or a single $-sigil embed; got {other:?} \
-             (expected $file / $base64 / $text)"
-        ),
-    };
-    Ok(Value::Array(bytes.into_iter().map(Value::from).collect()))
-}
-
-/// Reply-side mirror of [`resolve_bytes_params`] (issue 1944). The strict
-/// decoder emits a `Bytes` field as a JSON byte array; render it back to
-/// a bare string when the bytes are valid UTF-8 (the read-back-as-text
-/// ergonomic), else to `{"base64": …}`. Walks `schema` to reach a `Bytes`
-/// leaf nested in a composite, every other value untouched.
-fn render_bytes_reply(
-    value: serde_json::Value,
-    schema: &SchemaType,
-    inline_max: usize,
-) -> serde_json::Value {
-    use serde_json::Value;
-    match schema {
-        SchemaType::Bytes => render_bytes_leaf(value, inline_max),
-        SchemaType::Option(inner) => {
-            if value.is_null() {
-                value
-            } else {
-                render_bytes_reply(value, inner, inline_max)
-            }
-        }
-        SchemaType::Vec(inner) => match value {
-            Value::Array(items) => Value::Array(
-                items
-                    .into_iter()
-                    .map(|v| render_bytes_reply(v, inner, inline_max))
-                    .collect(),
-            ),
-            other => other,
-        },
-        SchemaType::Array { element, .. } => match value {
-            Value::Array(items) => Value::Array(
-                items
-                    .into_iter()
-                    .map(|v| render_bytes_reply(v, element, inline_max))
-                    .collect(),
-            ),
-            other => other,
-        },
-        SchemaType::Struct { fields, .. } => match value {
-            Value::Object(mut map) => {
-                for field in fields.iter() {
-                    if let Some(slot) = map.get_mut(&*field.name) {
-                        let taken = mem::take(slot);
-                        *slot = render_bytes_reply(taken, &field.ty, inline_max);
-                    }
-                }
-                Value::Object(map)
-            }
-            other => other,
-        },
-        SchemaType::Map {
-            value: value_schema,
-            ..
-        } => match value {
-            Value::Object(mut map) => {
-                for slot in map.values_mut() {
-                    let taken = mem::take(slot);
-                    *slot = render_bytes_reply(taken, value_schema, inline_max);
-                }
-                Value::Object(map)
-            }
-            other => other,
-        },
-        SchemaType::Enum { variants } => match value {
-            // Unit variant — bare string, no payload to descend into.
-            Value::String(_) => value,
-            // Non-unit variant — externally-tagged one-key object
-            // `{"VariantName": payload}`. Resolve the tag, then recurse
-            // into the payload according to the matched variant kind.
-            Value::Object(map) if map.len() == 1 => {
-                let (tag, payload) = map.into_iter().next().expect("len == 1");
-                let rendered = match variants.iter().find(|v| v.name() == tag.as_str()) {
-                    None | Some(EnumVariant::Unit { .. }) => payload,
-                    Some(EnumVariant::Tuple { fields, .. }) => match fields.as_ref() {
-                        [single] => render_bytes_reply(payload, single, inline_max),
-                        _ => match payload {
-                            Value::Array(items) => Value::Array(
-                                items
-                                    .into_iter()
-                                    .zip(fields.iter())
-                                    .map(|(v, f)| render_bytes_reply(v, f, inline_max))
-                                    .collect(),
-                            ),
-                            other => other,
-                        },
-                    },
-                    Some(EnumVariant::Struct { fields, .. }) => match payload {
-                        Value::Object(mut inner) => {
-                            for field in fields.iter() {
-                                if let Some(slot) = inner.get_mut(&*field.name) {
-                                    let taken = mem::take(slot);
-                                    *slot = render_bytes_reply(taken, &field.ty, inline_max);
-                                }
-                            }
-                            Value::Object(inner)
-                        }
-                        other => other,
-                    },
-                };
-                let mut out = serde_json::Map::new();
-                out.insert(tag, rendered);
-                Value::Object(out)
-            }
-            other => other,
-        },
-        _ => value,
-    }
-}
-
-/// Resolve the reply-`Bytes` spill threshold (`AETHER_MCP_REPLY_INLINE_MAX_BYTES`,
-/// default 256 KiB). A reply `Bytes` leaf larger than this is staged to a
-/// harness-host temp file and surfaced as `{"file": …}` instead of inlined as
-/// base64 (issue 2108) — the reply-side mirror of the input `{"$file": …}`
-/// embed. Deliberately distinct from the RPC frame cap: an over-frame reply
-/// never reaches this projection (the inbound frame decode fails first), and
-/// 64 MiB is far past the point base64 bloats the tool channel.
-fn reply_inline_max_bytes() -> usize {
-    use std::env;
-    const DEFAULT_REPLY_INLINE_MAX_BYTES: usize = 256 * 1024;
-    // Process-level tuning knob, not capability config: the out-of-process MCP
-    // front has no `Config`-derive plumbing, so it reads this directly the same
-    // way `main.rs` reads `AETHER_MCP_PORT`.
-    #[allow(clippy::disallowed_methods)]
-    let raw = env::var("AETHER_MCP_REPLY_INLINE_MAX_BYTES").ok();
-    raw.and_then(|s| s.trim().parse::<usize>().ok())
-        .unwrap_or(DEFAULT_REPLY_INLINE_MAX_BYTES)
-}
-
-/// Stage `bytes` to a unique harness-host temp file under `dir` and return its
-/// absolute path. Mirrors the boot-manifest staging writes (`stage_boot_manifest`):
-/// a blocking `std::fs::write` into the temp dir, named `aether-reply-<uuid>.bin`.
-fn spill_reply_bytes(dir: &Path, bytes: &[u8]) -> io::Result<PathBuf> {
-    // `std::fs`, not the module's `tokio::fs`: this write is blocking by design
-    // (the MCP front handling a tool reply is not latency-critical).
-    use std::fs as std_fs;
-    let path = dir.join(format!("aether-reply-{}.bin", Uuid::new_v4()));
-    std_fs::write(&path, bytes)?;
-    Ok(path)
-}
-
-/// Render one decoded `Bytes` value — a JSON array of byte numbers — for the
-/// caller. The ladder: a value over `inline_max` (regardless of UTF-8 validity)
-/// is spilled to a harness-host temp file and rendered as `{"file": …}`; a
-/// smaller value renders as a bare string (valid UTF-8) or a `{"base64": …}`
-/// object (binary). A value that isn't the array-of-bytes shape is returned
-/// untouched.
-fn render_bytes_leaf(value: serde_json::Value, inline_max: usize) -> serde_json::Value {
-    use std::env;
-    render_bytes_leaf_in(value, inline_max, &env::temp_dir())
-}
-
-/// `render_bytes_leaf` with an explicit spill directory — the seam the unit
-/// tests drive so spills land under a scratch dir rather than the real temp dir.
-fn render_bytes_leaf_in(
-    value: serde_json::Value,
-    inline_max: usize,
-    spill_dir: &Path,
-) -> serde_json::Value {
-    use serde_json::Value;
-    let Value::Array(items) = &value else {
-        return value;
-    };
-    let mut bytes = Vec::with_capacity(items.len());
-    for item in items {
-        match item.as_u64().and_then(|n| u8::try_from(n).ok()) {
-            Some(b) => bytes.push(b),
-            None => return value,
-        }
-    }
-    // Large reply (regardless of UTF-8 validity) → spill to a host temp file and
-    // surface a `{"file": …}` reference rather than inlining. On a spill IO error
-    // fall through to the in-band rendering below — the projector must never
-    // error or drop reply data (consistent with every other arm's pass-through).
-    if bytes.len() > inline_max
-        && let Ok(path) = spill_reply_bytes(spill_dir, &bytes)
-    {
-        let mut obj = serde_json::Map::new();
-        obj.insert(
-            "file".to_owned(),
-            Value::String(path.to_string_lossy().into_owned()),
-        );
-        return Value::Object(obj);
-    }
-    str::from_utf8(&bytes).map_or_else(
-        |_| {
-            let mut obj = serde_json::Map::new();
-            obj.insert("base64".to_owned(), Value::String(STANDARD.encode(&bytes)));
-            Value::Object(obj)
-        },
-        |s| Value::String(s.to_owned()),
-    )
-}
-
-/// Issue 963: render an `actor_logs` `LogTailResult::Err` into a
-/// tool-error message that names the agent-supplied mailbox, so an
-/// unregistered-mailbox query reads as "that mailbox doesn't exist"
-/// rather than a bare relayed substrate string. Factored out so the
-/// formatting is unit-testable without standing up a live engine.
-fn actor_logs_err_message(mailbox_name: &str, error: &str) -> String {
-    format!("actor_logs: mailbox \"{mailbox_name}\" — {error}")
-}
-
-/// Map ADR-0023 §4's level string to the `0..=4` byte the
-/// `aether.log.*` kinds carry. Case-insensitive. Returns an
-/// `invalid_params` error on unknown strings so a typoed `"Warn "`
-/// surfaces at the tool boundary rather than reaching the substrate.
-fn parse_level(s: &str) -> Result<u8, McpError> {
-    match s.to_ascii_lowercase().as_str() {
-        "trace" => Ok(0),
-        "debug" => Ok(1),
-        "info" => Ok(2),
-        "warn" => Ok(3),
-        "error" => Ok(4),
-        other => Err(McpError::invalid_params(
-            format!("unknown level {other:?}; expected trace|debug|info|warn|error"),
-            None,
-        )),
-    }
-}
-
-/// Inverse of [`parse_level`]: render the `0..=4` byte back to the
-/// canonical lowercase level string. Out-of-band bytes render as
-/// `"info"` (matches the existing fallback in
-/// `aether-capabilities::log`'s pre-issue-776 conversion).
-fn level_to_str(level: u8) -> &'static str {
-    match level {
-        0 => "trace",
-        1 => "debug",
-        3 => "warn",
-        4 => "error",
-        // 2 is "info"; out-of-band bytes also render as "info".
-        _ => "info",
-    }
-}
-
-// Imports for the `#[cfg(test)]` `RouteInventorySink` loopback fixture
-// (issue 2672). Brought into scope (rather than named by absolute path
-// inline) to satisfy the `clippy::absolute_paths` restriction.
-#[cfg(test)]
-use aether_actor::actor;
-#[cfg(test)]
-use aether_capabilities::engine::kinds::{CallSettled, RouteEnvelope};
-#[cfg(test)]
-use aether_substrate::actor::native::{NativeActor, NativeCtx, NativeInitCtx};
-#[cfg(test)]
-use aether_substrate::chassis::error::BootError;
-#[cfg(test)]
-use aether_substrate::mail::mailer::Mailer;
-#[cfg(test)]
-use std::sync::atomic::{AtomicUsize, Ordering};
-
-/// The canned live vocabulary a [`RouteInventorySink`] replies with, plus
-/// a counter of how many refresh RPCs it has fielded (issue 2672). Shared
-/// by value into the fixture so a test both controls the widened schema
-/// the refresh observes and asserts the refresh fired exactly once.
-#[cfg(test)]
-#[derive(Clone)]
-struct RouteLoopbackConfig {
-    reply: ListKindsResult,
-    calls: Arc<AtomicUsize>,
-}
-
-/// `#[cfg(test)]` loopback engines-cap double (issue 2672). Registers at
-/// the `aether.engine` mailbox — the id the `RpcServerCapability` routes
-/// every `engine = Some` `Call` to via a `RouteEnvelope` — and answers the
-/// harness's `aether.inventory.kinds` refresh RPC locally with a canned
-/// [`ListKindsResult`](aether_kinds::ListKindsResult), so the
-/// field-mismatch refresh-and-retry path in [`Mcp::resolve_and_encode`] is
-/// exercised end-to-end without forking a real substrate + proxy.
-///
-/// Lives at file root (not nested in `mod tests`) so the `#[actor]`
-/// macro's marker emission stays addressable, mirroring the engines-cap's
-/// own `ReplySink`. It stands in for the real `EngineServer` (never
-/// co-installed with it, so the shared `aether.engine` mailbox id is
-/// unambiguous): on a `RouteEnvelope` it pushes the reply and the
-/// `CallSettled` terminal straight back to the originating server,
-/// correlation preserved, so the forwarded wire call closes the way a
-/// proxy's `CallSettled` would.
-#[cfg(test)]
-struct RouteInventorySink {
-    reply: ListKindsResult,
-    calls: Arc<AtomicUsize>,
-    mailer: Arc<Mailer>,
-}
-
-#[cfg(test)]
-#[actor(singleton)]
-impl NativeActor for RouteInventorySink {
-    type Config = RouteLoopbackConfig;
-    const NAMESPACE: &'static str = "aether.engine";
-
-    fn init(config: RouteLoopbackConfig, ctx: &mut NativeInitCtx<'_>) -> Result<Self, BootError> {
-        Ok(Self {
-            reply: config.reply,
-            calls: config.calls,
-            // Cached like the real engines cap does (its `on_route`
-            // propagates the inbound reply-to, which `NativeCtx` sends
-            // would overwrite with this cap as sender).
-            mailer: ctx.mailer(),
-        })
-    }
-
-    #[handler::single]
-    fn on_route(&mut self, ctx: &mut NativeCtx<'_>, _mail: RouteEnvelope) {
-        use aether_substrate::mail::{Mail, Source, SourceAddr};
-
-        self.calls.fetch_add(1, Ordering::Relaxed);
-        let reply_to = ctx.reply_target();
-        // A routed call always carries a Component reply-to (the
-        // originating server); without one there's nowhere to stream to.
-        let SourceAddr::Component(target) = reply_to.addr else {
-            return;
-        };
-        let correlation = reply_to.correlation_id;
-
-        // ReplyEvent: the canned live vocabulary. The server matches it to
-        // the in-flight wire call by the preserved correlation.
-        self.mailer.push(
-            Mail::new(
-                target,
-                <ListKindsResult as Kind>::ID,
-                self.reply.encode_into_bytes(),
-                1,
-            )
-            .with_reply_to(Source::with_correlation(SourceAddr::None, correlation)),
-        );
-        // ReplyEnd: a forwarded call has no local chain to settle, so the
-        // server's `engine = Some` path waits on this explicit terminal
-        // (in production the proxy lifts the substrate's `ReplyEnd` into
-        // it). Pushed after the reply so the server writes the ReplyEvent
-        // frame first, then closes on the CallSettled.
-        self.mailer.push(
-            Mail::new(
-                target,
-                <CallSettled as Kind>::ID,
-                CallSettled::Ok.encode_into_bytes(),
-                1,
-            )
-            .with_reply_to(Source::with_correlation(SourceAddr::None, correlation)),
-        );
     }
 }
 
@@ -3257,6 +1453,7 @@ mod tests {
     use aether_substrate::mail::registry::Registry;
     use std::path::PathBuf;
     use std::process;
+    use std::sync::atomic::{AtomicUsize, Ordering};
     use std::time::{SystemTime, UNIX_EPOCH};
     use std::{env as std_env, fs as std_fs};
